chore: Move benchmarks to a test file
Note these don't use #[bench] attribute as this currently works on rust
nightly.
